+++ seed/java-sdk/respect-optional-request-body/src/main/java/com/snippets/Example1.java
@@ -1,0 +1,18 @@
+package com.snippets;
+
+import com.seed.api.SeedApiClient;
+import com.seed.api.requests.RefundBody;
+import com.seed.api.types.RefundRequest;
+
+public class Example1 {
+    public static void main(String[] args) {
+        SeedApiClient client =
+                SeedApiClient.builder().url("https://api.fern.com").build();
+
+        client.refund(
+                "id",
+                RefundBody.builder()
+                        .body(RefundRequest.builder().amount(1.1).build())
+                        .build());
+    }
+}