+++ seed/java-sdk/respect-optional-request-body/src/main/java/com/snippets/Example3.java
@@ -1,0 +1,18 @@
+package com.snippets;
+
+import com.seed.api.SeedApiClient;
+import com.seed.api.requests.RequiredRefundRequest;
+import com.seed.api.types.RefundRequest;
+
+public class Example3 {
+    public static void main(String[] args) {
+        SeedApiClient client =
+                SeedApiClient.builder().url("https://api.fern.com").build();
+
+        client.requiredRefund(
+                "id",
+                RequiredRefundRequest.builder()
+                        .body(RefundRequest.builder().amount(1.1).build())
+                        .build());
+    }
+}