+++ seed/java-sdk/respect-optional-request-body/src/main/java/com/snippets/Example5.java
@@ -1,0 +1,13 @@
+package com.snippets;
+
+import com.seed.api.SeedApiClient;
+import com.seed.api.types.RefundRequest;
+
+public class Example5 {
+    public static void main(String[] args) {
+        SeedApiClient client =
+                SeedApiClient.builder().url("https://api.fern.com").build();
+
+        client.bulkRefund(RefundRequest.builder().amount(1.1).build());
+    }
+}