+++ seed/python-sdk/respect-optional-request-body/src/seed/client.py
@@ -7,7 +7,6 @@ from .core.client_wrapper import AsyncClientWrapper, SyncClientWrapper
 from .core.logging import LogConfig, Logger
 from .core.request_options import RequestOptions
 from .raw_client import AsyncRawSeedApi, RawSeedApi
-from .types.refund_request import RefundRequest
 
 # this is used as the default value for optional parameters
 OMIT = typing.cast(typing.Any, ...)
@@ -98,18 +97,14 @@ class SeedApi:
         return self._raw_client
 
     def refund(
-        self,
-        id: str,
-        *,
-        request: typing.Optional[RefundRequest] = None,
-        request_options: typing.Optional[RequestOptions] = None,
+        self, id: str, *, amount: typing.Optional[float] = OMIT, request_options: typing.Optional[RequestOptions] = None
     ) -> None:
         """
         Parameters
         ----------
         id : str
 
-        request : typing.Optional[RefundRequest]
+        amount : typing.Optional[float]
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -127,9 +122,10 @@ class SeedApi:
         )
         client.refund(
             id="refund-id",
+            amount=60.0,
         )
         """
-        _response = self._raw_client.refund(id, request=request, request_options=request_options)
+        _response = self._raw_client.refund(id, amount=amount, request_options=request_options)
         return _response.data
 
     def required_refund(
@@ -165,12 +161,12 @@ class SeedApi:
         return _response.data
 
     def bulk_refund(
-        self, *, request: typing.Optional[RefundRequest] = None, request_options: typing.Optional[RequestOptions] = None
+        self, *, amount: typing.Optional[float] = OMIT, request_options: typing.Optional[RequestOptions] = None
     ) -> None:
         """
         Parameters
         ----------
-        request : typing.Optional[RefundRequest]
+        amount : typing.Optional[float]
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -181,16 +177,14 @@ class SeedApi:
 
         Examples
         --------
-        from seed import RefundRequest, SeedApi
+        from seed import SeedApi
 
         client = SeedApi(
             base_url="https://yourhost.com/path/to/api",
         )
-        client.bulk_refund(
-            request=RefundRequest(),
-        )
+        client.bulk_refund()
         """
-        _response = self._raw_client.bulk_refund(request=request, request_options=request_options)
+        _response = self._raw_client.bulk_refund(amount=amount, request_options=request_options)
         return _response.data
 
 
@@ -295,18 +289,14 @@ class AsyncSeedApi:
         return self._raw_client
 
     async def refund(
-        self,
-        id: str,
-        *,
-        request: typing.Optional[RefundRequest] = None,
-        request_options: typing.Optional[RequestOptions] = None,
+        self, id: str, *, amount: typing.Optional[float] = OMIT, request_options: typing.Optional[RequestOptions] = None
     ) -> None:
         """
         Parameters
         ----------
         id : str
 
-        request : typing.Optional[RefundRequest]
+        amount : typing.Optional[float]
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -329,12 +319,13 @@ class AsyncSeedApi:
         async def main() -> None:
             await client.refund(
                 id="refund-id",
+                amount=60.0,
             )
 
 
         asyncio.run(main())
         """
-        _response = await self._raw_client.refund(id, request=request, request_options=request_options)
+        _response = await self._raw_client.refund(id, amount=amount, request_options=request_options)
         return _response.data
 
     async def required_refund(
@@ -378,12 +369,12 @@ class AsyncSeedApi:
         return _response.data
 
     async def bulk_refund(
-        self, *, request: typing.Optional[RefundRequest] = None, request_options: typing.Optional[RequestOptions] = None
+        self, *, amount: typing.Optional[float] = OMIT, request_options: typing.Optional[RequestOptions] = None
     ) -> None:
         """
         Parameters
         ----------
-        request : typing.Optional[RefundRequest]
+        amount : typing.Optional[float]
 
         request_options : typing.Optional[RequestOptions]
             Request-specific configuration.
@@ -396,7 +387,7 @@ class AsyncSeedApi:
         --------
         import asyncio
 
-        from seed import AsyncSeedApi, RefundRequest
+        from seed import AsyncSeedApi
 
         client = AsyncSeedApi(
             base_url="https://yourhost.com/path/to/api",
@@ -404,12 +395,10 @@ class AsyncSeedApi:
 
 
         async def main() -> None:
-            await client.bulk_refund(
-                request=RefundRequest(),
-            )
+            await client.bulk_refund()
 
 
         asyncio.run(main())
         """
-        _response = await self._raw_client.bulk_refund(request=request, request_options=request_options)
+        _response = await self._raw_client.bulk_refund(amount=amount, request_options=request_options)
         return _response.data