+++ seed/rust-sdk/respect-optional-request-body/dynamic-snippets/example1.rs
@@ -9,11 +9,11 @@ async fn main() {
     let client = ApiClient::new(config).expect("Failed to build client");
     client
         .refund(
-            &"refund-id".to_string(),
-            &Some(RefundRequest {
-                amount: Some(60.0),
+            &"id".to_string(),
+            &RefundRequest {
+                amount: Some(1.1),
                 ..Default::default()
-            }),
+            },
             None,
         )
         .await;