+++ seed/rust-sdk/respect-optional-request-body/dynamic-snippets/example4.rs
@@ -8,10 +8,8 @@ async fn main() {
     };
     let client = ApiClient::new(config).expect("Failed to build client");
     client
-        .required_refund(
-            &"id".to_string(),
+        .bulk_refund(
             &RefundRequest {
-                amount: Some(1.1),
                 ..Default::default()
             },
             None,